Fix bugs, clean up LOCK_SWAP, remove delegatecall functionality
Re delegatecall: does not work well with the new Core contract not allowing to specify the owner of the saved balance.
Also, may be replaced by mechanism in the future in which the router can accept forwarded locks and does not have to care about settling balances
(thus, eliminating the main benefit of not requiring approvals for delegatecalls).